add fake initial data, but haven't fixed problem yet

+++ src/main/java/no/kristiania/services/CustomerService.java
@@ -1,6 +1,9 @@
 package no.kristiania.services;
 
+import no.kristiania.errors.CustomerNotFoundException;
+import no.kristiania.errors.OrderNotFoundException;
 import no.kristiania.repositories.customers.Customer;
+import no.kristiania.repositories.customers.CustomerAddressRepo;
 import no.kristiania.repositories.customers.CustomerRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,10 +13,15 @@ import java.util.List;
 @Service
 public class CustomerService {
     private final CustomerRepo customerRepo;
-
+    private final CustomerAddressRepo customerAddressRepo;
    @Autowired
-    public CustomerService(CustomerRepo customerRepo) {
-        this.customerRepo = customerRepo;
+    public CustomerService(
+            CustomerRepo customerRepo,
+            CustomerAddressRepo customerAddressRepo
+           ) {
+
+       this.customerRepo = customerRepo;
+       this.customerAddressRepo = customerAddressRepo;
     }
 
     // TODO: Fetching a customer should show their addresses and order history.
@@ -24,15 +32,24 @@ public class CustomerService {
     }
 
     public Customer createCustomer(Customer customer) {
-        return customerRepo.save(customer);
+       return customerRepo.save(customer);
     }
 
     public Customer getCustomerById(Long id){
-        return customerRepo.findById(id).orElse(null);
+       return customerRepo.findById(id).orElse(null);
     }
 
     public void deleteCustomer(Long id) {
-        customerRepo.deleteById(id);
+        if (!customerRepo.existsById(id)) {
+            throw new CustomerNotFoundException("Customer not found for ID: " + id);
+        }
+
+       customerRepo.deleteById(id);
+    }
+
+    public void deleteAllCustomers() {
+       customerAddressRepo.deleteAll();
+       customerRepo.deleteAll();
     }
 
 }
